requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
beautifulsoup4==4.11.1
    # via -r requirements.in
boto3==1.23.5
    # via -r requirements.in
botocore==1.26.5
    # via
    #   boto3
    #   s3transfer
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
dropbox==11.31.0
    # via -r requirements.in
idna==3.3
    # via requests
imutils==0.5.4
    # via -r requirements.in
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
json-minify==0.3.0
    # via -r requirements.in
labelimg==1.8.6
    # via -r requirements.in
lxml==4.8.0
    # via
    #   -r requirements.in
    #   labelimg
numpy==1.22.4
    # via
    #   opencv-contrib-python
    #   opencv-python
opencv-contrib-python==4.5.5.64
    # via -r requirements.in
opencv-python==4.5.5.64
    # via -r requirements.in
ply==3.11
    # via stone
pyqt5==5.15.7
    # via labelimg
pyqt5-qt5==5.15.2
    # via pyqt5
pyqt5-sip==12.11.0
    # via pyqt5
python-dateutil==2.8.2
    # via botocore
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.27.1
    # via dropbox
s3transfer==0.5.2
    # via boto3
six==1.16.0
    # via
    #   dropbox
    #   python-dateutil
    #   stone
soupsieve==2.3.2.post1
    # via beautifulsoup4
stone==3.3.1
    # via dropbox
urllib3==1.26.9
    # via
    #   botocore
    #   requests
